pyproject.toml
[project]
name = "Streamlit-Chatbot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Alex Breault", email = "bigalbreault@gmail.com"},
]
dependencies = [
    "streamlit>=1.43.1",
    "PyPDF2>=3.0.1",
    "openai>=1.65.4",
    "watchdog>=6.0.0",
    "aws-cdk-lib>=2.182.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
